feat: compile base rollup as a circuit (#3739)
This PR adds a variant of the base rollup that is a circuit. This makes
sure that PRs don't break compilation as a circuit of the base rollup
and tracks its constraint count.
After the last noir pull, thanks to @TomAFrench compilation is now much
faster, allowing to compile the base rollup as a circuit in a reasonable
time (prev. it took 5 minutes, now some seconds)